requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
psycopg2-binary
sqlalchemy
jupyter
joblib